Change MojoHandle to 64 bits
We no longer have need to preserve Mojo Core ABI stability because
consuming Mojo Core as a shared library is no longer supported. This
changes MojoHandle to a 64-bit value in preparation for the
introduction of ipcz -- which itself uses 64-bit handles -- making it
trivial to reuse existing Mojo C++ handle types to wrap Mojo and ipcz
handles alike.

Note that this does not impact the wire format in any way, because
handle values are only for the (process-local) API interface. In
messages, handles are still encoded as 32-bit indices to some
object metadata within the message.

Bug: 1299283
Binary-Size: This intentionally increases the size of a widely used data type from 32 bits to 64 bits, a change which is required to support other improvements. The 110 kB binary size increase is a direct result of that, and it does not seem to be avoidable.
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3493659
Commit-Queue: Ken Rockot <rockot@google.com>
Cr-Commit-Position: refs/heads/main@{#976468}

+++ content/browser/android/dialog_overlay_impl.cc
@@ -305,7 +305,7 @@ class AndroidOverlaySyncHelper {
 
 static void JNI_DialogOverlayImpl_NotifyDestroyedSynchronously(
     JNIEnv* env,
-    int message_pipe_handle) {
+    jlong message_pipe_handle) {
   mojo::MessagePipeHandle handle(message_pipe_handle);
   mojo::ScopedMessagePipeHandle scoped_handle(handle);
   mojo::Remote<media::mojom::AndroidOverlayClient> remote(
