+++ content/public/android/java/src/org/chromium/content/browser/InterfaceRegistrarImpl.java
@@ -23,7 +23,7 @@ class InterfaceRegistrarImpl {
     private static boolean sHasRegisteredRegistrars;
 
     @CalledByNative
-    static void createInterfaceRegistry(int nativeHandle) {
+    static void createInterfaceRegistry(long nativeHandle) {
         ensureSingletonRegistrarsAreRegistered();
 
         InterfaceRegistry registry = InterfaceRegistry.create(
@@ -32,14 +32,14 @@ class InterfaceRegistrarImpl {
     }
 
     @CalledByNative
-    static void createInterfaceRegistryOnIOThread(int nativeHandle) {
+    static void createInterfaceRegistryOnIOThread(long nativeHandle) {
         InterfaceRegistry registry = InterfaceRegistry.create(
                 CoreImpl.getInstance().acquireNativeHandle(nativeHandle).toMessagePipeHandle());
         registerInterfacesOnIOThread(registry);
     }
 
     @CalledByNative
-    static void createInterfaceRegistryForWebContents(int nativeHandle, WebContents webContents) {
+    static void createInterfaceRegistryForWebContents(long nativeHandle, WebContents webContents) {
         ensureSingletonRegistrarsAreRegistered();
 
         InterfaceRegistry registry = InterfaceRegistry.create(
@@ -49,7 +49,7 @@ class InterfaceRegistrarImpl {
 
     @CalledByNative
     static void createInterfaceRegistryForRenderFrameHost(
-            int nativeHandle, RenderFrameHost renderFrameHost) {
+            long nativeHandle, RenderFrameHost renderFrameHost) {
         ensureSingletonRegistrarsAreRegistered();
 
         InterfaceRegistry registry = InterfaceRegistry.create(
