+++ services/device/android/java/src/org/chromium/services/device/InterfaceRegistrar.java
@@ -19,8 +19,7 @@ import org.chromium.services.service_manager.InterfaceRegistry;
 @JNINamespace("device")
 class InterfaceRegistrar {
     @CalledByNative
-    static void createInterfaceRegistryForContext(
-            int nativeHandle, NfcDelegate nfcDelegate) {
+    static void createInterfaceRegistryForContext(long nativeHandle, NfcDelegate nfcDelegate) {
         // Note: The bindings code manages the lifetime of this object, so it
         // is not necessary to hold on to a reference to it explicitly.
         InterfaceRegistry registry = InterfaceRegistry.create(
